[Script] Move unicode data files in to a separate module, so the model stuff is not Foundation (i.e. OS) dependent.
