+++ UnicodeData/Sources/GenerateUnicodeData/main.swift
@@ -12,9 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import Foundation
+import UnicodeDataFiles
 import UnicodeDataTools
 
-var output =
+#if true  // When debugging, set to false to output files to the tmp directory.
+  let _outputDir = CommandLine.arguments.dropFirst().first ?? ""
+  precondition(!_outputDir.isEmpty, "Missing command-line argument: output directory")
+  let outputDir = URL(fileURLWithPath: _outputDir, isDirectory: true)
+#else
+  let outputDir = URL(fileURLWithPath: NSTemporaryDirectory(), isDirectory: true)
+    .appendingPathComponent("IDNA", isDirectory: true)
+  try FileManager.default.createDirectory(at: outputDir, withIntermediateDirectories: true, attributes: nil)
+#endif
+
+
+let header =
   #"""
   // Copyright The swift-url Contributors.
   //
@@ -36,11 +49,25 @@ var output =
   // Generated by GenerateUnicodeData
 
   """# + "\n"
-//output += try generateIDNAMappingData()
-output += try generateIDNAValidationData()
 
-// Fix up trailing newlines.
-precondition(output.last == "\n")
-output.removeLast()
-// Print.
-print(output)
+// Mapping data.
+let mappingDataFile = try String(contentsOf: URLOfUnicodeDataFile(named: "IdnaMappingTable")!)
+let mappingDB = IDNAMappingDatabase(parsing: mappingDataFile)
+do {
+  let swiftSourceFile = header + mappingDB.printAsSwiftSourceCode(name: "_idna")
+  try Data(swiftSourceFile.utf8).write(to: outputDir.appendingPathComponent("MappingData.swift", isDirectory: false))
+}
+
+// Validation data.
+do {
+  let derivedBidiClassTxt = try String(contentsOf: URLOfUnicodeDataFile(named: "DerivedBidiClass")!)
+  let derivedJoiningTypeTxt = try String(contentsOf: URLOfUnicodeDataFile(named: "DerivedJoiningType")!)
+  let db = IDNAValidationDatabase(
+    mappingDB: mappingDB,
+    derivedBidiClassTxt: derivedBidiClassTxt,
+    derivedJoiningTypeTxt: derivedJoiningTypeTxt
+  )
+  let swiftSourceFile = header + db.printAsSwiftSourceCode(name: "_bidi")
+  try Data(swiftSourceFile.utf8).write(to: outputDir.appendingPathComponent("ValidationData.swift", isDirectory: false))
+}
+print("Files generated in: \(outputDir)")
