+++ UnicodeData/Sources/UnicodeDataFiles/TableDefinitions/DerivedBidiClass.txt
