+++ UnicodeData/Sources/UnicodeDataFiles/TableDefinitions/DerivedJoiningType.txt
