+++ UnicodeData/Sources/UnicodeDataFiles/TableDefinitions/IdnaMappingTable.txt
